refactor: refactor create petition
