+++ src/components/organisms/Header/consts.ts
@@ -1,0 +1,4 @@
+export const links = [
+  { to: "/", text: "Главная" },
+  { to: "/tickets", text: "История" },
+];
